+++ developer_resources/sample_import_node.py
@@ -1,0 +1,17 @@
+Import(
+    names=[
+        ImportAlias(
+            name=Name(
+                value="os",
+                lpar=[],
+                rpar=[],
+            ),
+            asname=None,
+            comma=MaybeSentinel.DEFAULT,
+        ),
+    ],
+    semicolon=MaybeSentinel.DEFAULT,
+    whitespace_after_import=SimpleWhitespace(
+        value=" ",
+    ),
+)