+++ tests/e2e_tests/test_basic_behavior.py
@@ -107,7 +107,7 @@ def test_default_one_error(tmp_path_factory, e2e_config):
     stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
     assert "Traceback (most recent call last)" in stderr
     assert 'system_info_script.py", line 4, in <module>' in stderr
-    assert "ModuleNotFoundError: No module named 'o'" in stderr
+    assert "ModuleNotFoundError: No module named " in stderr
 
     # Read modified file; should have changed only one import statement.
     modified_source = path_dst.read_text()
@@ -139,7 +139,7 @@ def test_two_bugs(tmp_path_factory, e2e_config):
     stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
     assert "Traceback (most recent call last)" in stderr
     assert 'system_info_script.py", line 3, in <module>' in stderr
-    assert "ModuleNotFoundError: No module named 'ys'" in stderr
+    assert "ModuleNotFoundError: No module named " in stderr
 
     # Read modified file; should have changed both import statements.
     modified_source = path_dst.read_text()
@@ -173,11 +173,23 @@ def test_random_import_affected(tmp_path_factory, e2e_config):
     stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
     assert "Traceback (most recent call last)" in stderr
     assert 'ten_imports.py", line 6, in <module>' in stderr
-    assert "ModuleNotFoundError: No module named 'clendar'" in stderr
+    assert "ModuleNotFoundError: No module named " in stderr
 
     # Read modified file; should have changed import statement.
     modified_source = path_dst.read_text()
-    assert "import clendar" in modified_source
+    assert "import calendar" not in modified_source
+    pkgs = [
+        "os",
+        "sys",
+        "re",
+        "random",
+        "difflib",
+        "zoneinfo",
+        "array",
+        "pprint",
+        "enum",
+    ]
+    assert all([p in modified_source for p in pkgs])
 
 
 def test_random_py_file_affected(tmp_path_factory, e2e_config):
@@ -209,7 +221,7 @@ def test_random_py_file_affected(tmp_path_factory, e2e_config):
     stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
     assert "Traceback (most recent call last)" in stderr
     assert 'ten_imports.py", line 7, in <module>' in stderr
-    assert "ModuleNotFoundError: No module named 'zoneino'" in stderr
+    assert "ModuleNotFoundError: No module named " in stderr
 
     # Other file should not be changed.
     assert filecmp.cmp(e2e_config.path_system_info, path_dst_system_info)
@@ -294,10 +306,11 @@ def test_target_file(tmp_path_factory, e2e_config):
     stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
     assert "Traceback (most recent call last)" in stderr
     assert 'system_info_script.py", line 4, in <module>' in stderr
-    assert "ModuleNotFoundError: No module named 'o'" in stderr
+    assert "ModuleNotFoundError: No module named " in stderr
 
     # Other file should not be changed.
     assert filecmp.cmp(e2e_config.path_ten_imports, path_dst_ten_imports)
+
 
 def test_attribute_error(tmp_path_factory, e2e_config):
     """py-bugger --exception-type AttributeError"""
@@ -310,7 +323,10 @@ def test_attribute_error(tmp_path_factory, e2e_config):
     shutil.copyfile(e2e_config.path_name_picker, path_dst)
 
     # Run py-bugger against directory.
-    cmd = f"py-bugger --exception-type AttributeError --target-dir {tmp_path.as_posix()}"
+    cmd = (
+        f"py-bugger --exception-type AttributeError --target-dir {tmp_path.as_posix()}"
+    )
+    print("cmd:", cmd)
     cmd_parts = shlex.split(cmd)
 
     stdout = subprocess.run(cmd_parts, capture_output=True).stdout.decode()
@@ -323,11 +339,13 @@ def test_attribute_error(tmp_path_factory, e2e_config):
     stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
     assert "Traceback (most recent call last)" in stderr
     assert 'name_picker.py", line 7, in <module>' in stderr
-    assert "AttributeError: 'str' object has no attribute 'tite'. Did you mean: 'title'?" in stderr
+    assert "AttributeError: 'str' object has no attribute " in stderr
+    assert "Did you mean: " in stderr
 
     # Make sure only one attribute was affected.
     modified_source = path_dst.read_text()
     assert "random.choice(names)" in modified_source
+
 
 def test_one_node_changed(tmp_path_factory, e2e_config):
     """Test that only one node in a file is modified for identical nodes."""
@@ -339,7 +357,9 @@ def test_one_node_changed(tmp_path_factory, e2e_config):
     shutil.copyfile(e2e_config.path_dog, path_dst)
 
     # Run py-bugger against directory.
-    cmd = f"py-bugger --exception-type AttributeError --target-dir {tmp_path.as_posix()}"
+    cmd = (
+        f"py-bugger --exception-type AttributeError --target-dir {tmp_path.as_posix()}"
+    )
     cmd_parts = shlex.split(cmd)
 
     stdout = subprocess.run(cmd_parts, capture_output=True).stdout.decode()
@@ -351,10 +371,44 @@ def test_one_node_changed(tmp_path_factory, e2e_config):
     cmd_parts = shlex.split(cmd)
     stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
     assert "Traceback (most recent call last)" in stderr
-    assert 'dog.py", line 9, in <module>' in stderr
-    assert "AttributeError: 'Dog' object has no attribute 'name'. Did you mean: 'nam'?" in stderr
+    assert 'dog.py", line 10, in <module>' in stderr
+    assert "AttributeError: 'Dog' object has no attribute " in stderr
+    assert "Did you mean: " in stderr
 
     # Make sure only one attribute was affected.
     modified_source = path_dst.read_text()
     assert "self.name" in modified_source
     assert "self.nam" in modified_source
+
+
+def test_random_node_changed(tmp_path_factory, e2e_config):
+    """Test that a random node in a file is modified if it has numerous identical nodes."""
+    # Copy sample code to tmp dir.
+    tmp_path = tmp_path_factory.mktemp("sample_code")
+    print(f"\nCopying code to: {tmp_path.as_posix()}")
+
+    path_dst = tmp_path / e2e_config.path_identical_attributes.name
+    shutil.copyfile(e2e_config.path_identical_attributes, path_dst)
+
+    # Run py-bugger against directory.
+    cmd = (
+        f"py-bugger --exception-type AttributeError --target-dir {tmp_path.as_posix()}"
+    )
+    cmd_parts = shlex.split(cmd)
+
+    stdout = subprocess.run(cmd_parts, capture_output=True).stdout.decode()
+
+    assert "All requested bugs inserted." in stdout
+
+    # Run file, should raise AttributeError.
+    cmd = f"{e2e_config.python_cmd.as_posix()} {path_dst.as_posix()}"
+    cmd_parts = shlex.split(cmd)
+    stderr = subprocess.run(cmd_parts, capture_output=True).stderr.decode()
+    assert "Traceback (most recent call last)" in stderr
+    assert 'identical_attributes.py", line 18, in <module>' in stderr
+    assert "AttributeError: module 'random' has no attribute " in stderr
+    assert "Did you mean: " in stderr
+
+    # Make sure only one attribute was affected.
+    modified_source = path_dst.read_text()
+    assert modified_source.count("random.choice(") == 19