+++ tests/sample_code/sample_scripts/dog.py
@@ -5,5 +5,6 @@ class Dog:
     def say_hi(self):
         print(f"Hi, I'm {self.name} the dog!")
 
+
 dog = Dog("Willie")
 dog.say_hi()