+++ tests/sample_code/sample_scripts/identical_attributes.py
@@ -1,0 +1,33 @@
+import random
+
+numbers = [1, 2, 3, 4, 5]
+
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+random.choice(numbers)
+
+random.choices(numbers, k=3)
+random.choices(numbers, k=3)
+random.choices(numbers, k=3)
+random.choices(numbers, k=3)
+random.choices(numbers, k=3)
+random.choices(numbers, k=3)
+random.choices(numbers, k=3)
+random.choices(numbers, k=3)